+++ ltnvm/objects/FxPointer.hxx
@@ -6,8 +6,8 @@
 namespace ltn::vm {
 	struct FxPointer {
 		const std::uint8_t * ptr;
-		std::size_t params;
-		std::size_t arity() const {
+		std::uint64_t params;
+		std::uint64_t arity() const {
 			return this->params;
 		}
 		std::vector<Value> captured;
