+++ ltnvm/objects/iter/StringIterator.cxx
@@ -20,7 +20,7 @@ namespace ltn::vm {
 	Value StringIterator::get(Heap & heap) {
 		auto & arr = heap.read<String>(this->ref);
 		if(this->index < std::ssize(arr) && this->index >= 0) {
-			return value::character(arr[this->index]);
+			return value::character(arr[static_cast<std::uint64_t>(this->index)]);
 		}
 		else {
 			return value::iterator_stop;
